==> notebooks/02_generate_embeddings.ipynb <==
import pandas as pd
import numpy as np
from sentence_transformers import SentenceTransformer

# Load the cleaned Spotify dataset for embedding
df = pd.read_csv('../data/SpotifyFeatures_cleaned.csv')
print("Shape of data:", df.shape)
df['combined_text'].head()

# Load a pre-trained SBERT model
model = SentenceTransformer('all-MiniLM-L6-v2')  # This model is fast and accurate for semantic similarity


# Create a list of sentences from the combined_text column
sentences = df['combined_text'].tolist()

# Generate SBERT embeddings for each song
embeddings = model.encode(sentences, batch_size=64, show_progress_bar=True)


# Save the embeddings as a NumPy array for future use
np.save('../data/song_embeddings.npy', embeddings)
print("Embeddings saved! Shape:", embeddings.shape)


# Inspect the first embedding to ensure correct dimensionality (should be 384)
print("First embedding:", embeddings[0])
print("Embedding dimension:", len(embeddings[0]))




embeddings = np.load('../data/song_embeddings.npy')

# Take a random sample of N songs (e.g., 1000)
N = 15000

df_sample = df.sample(n=N, random_state=42)
sample_indices = df_sample.index.tolist()

# Select only those embeddings
embeddings_sample = embeddings[sample_indices]

# Save the sampled DataFrame and .npy file
df_sample.to_csv('../data/SpotifyFeatures_sample.csv', index=False)
np.save('../data/song_embeddings_sample.npy', embeddings_sample)
